slash/app/requirements.txt
flask
pwntools